requirements.txt
flask
flask-restful
flask-jwt-extended
python-dotenv
marshmallow-sqlalchemy
flask-marshmallow
flask-sqlalchemy
requests
flask-mail
newsapi-python
psutil
psycopg2-binary
schedule
flask-testing
pytest
